+++ wagtailgeowidget/static/wagtailgeowidget/js/geocoder-field.js
@@ -136,6 +136,12 @@ NominatimGeocoderField.prototype.geocodeSearch = function (query) {
             self.field.trigger("searchGeocoded", [
                 { lat: location.lat, lng: location.lon },
             ]);
+        })
+        .catch((error) => {
+            self.displayWarning("Nominatim Error: " + error, {
+                field: self.field,
+            });
+            throw error;
         });
 };
 
@@ -234,6 +240,14 @@ MapboxGeocoderField.prototype.geocodeSearch = function (query) {
     fetch(url)
         .then((response) => response.json())
         .then((data) => {
+            // https://docs.mapbox.com/api/search/geocoding/#geocoding-api-errors
+            if (data.message) {
+                self.displayWarning("Map box Error: " + data.message, {
+                    field: self.field,
+                });
+                return;
+            }
+
             if (!data.features.length) {
                 self.displayWarning(
                     self.translations.error_could_not_geocode_address.replace(
@@ -256,5 +270,11 @@ MapboxGeocoderField.prototype.geocodeSearch = function (query) {
             self.field.trigger("searchGeocoded", [
                 { lat: coordinates[1], lng: coordinates[0] },
             ]);
+        })
+        .catch((error) => {
+            self.displayWarning("Map box Error: " + error, {
+                field: self.field,
+            });
+            throw error;
         });
 };